+++ chassis.go
@@ -42,8 +42,8 @@ import (
 	"github.com/go-chassis/go-chassis/core/tracing"
 	"github.com/go-chassis/go-chassis/eventlistener"
 
-	// metric
-	_ "github.com/go-chassis/go-chassis/metrics/prom"
+	// prometheus reporter for circuit breaker metrics
+	_ "github.com/go-chassis/go-chassis/third_party/forked/afex/hystrix-go/hystrix/reporter"
 
 	// aes package handles security related plugins
 	_ "github.com/go-chassis/go-chassis/security/plugins/aes"
@@ -56,9 +56,9 @@ import (
 	"github.com/go-chassis/go-chassis/configcenter"
 	"github.com/go-chassis/go-chassis/control"
 	"github.com/go-chassis/go-chassis/core/metadata"
-	"github.com/go-chassis/go-chassis/metrics"
 	"github.com/go-chassis/go-chassis/pkg/circuit"
 	"github.com/go-chassis/go-chassis/pkg/runtime"
+	"github.com/go-chassis/go-chassis/third_party/forked/afex/hystrix-go/hystrix"
 	"github.com/go-mesh/openlogging"
 )
 
@@ -176,9 +176,7 @@ func (c *chassis) initialize() error {
 	if err = tracing.Init(); err != nil {
 		return err
 	}
-	if err = metrics.Init(); err != nil {
-		return err
-	}
+	go hystrix.StartReporter()
 	circuit.Init()
 	eventlistener.Init()
 	c.Initialized = true
