+++ core/loadbalancer/struct_test.go
@@ -2,7 +2,7 @@ package loadbalancer_test
 
 import (
 	"github.com/go-chassis/go-chassis/core/loadbalancer"
-	"gopkg.in/go-playground/assert.v1"
+	"github.com/stretchr/testify/assert"
 	"testing"
 	"time"
 )
